#162: RecordStore/Lucene based FeatureStore (http://polymap.org/polymap3/ticket/162)   - NEW: catalog integration of the RDataStore     - H2 removed from toolbar     - new action that creates a new RDataStore in catalog added   - NEW: LuceneQueryDialect: code added that converts OGC filter to Lucene query   - NEW: RDataStore: basic support for SimpleFeature added   - GeometryValueCoder: use WKB as default   - LuceneRecordStore: use mmap on 32bit Linux if index < 500MB   - ServiceTypeFolder: use IService#getShortName() as fallback for folder name   - CatalogView: import action tooltip

+++ plugins/org.polymap.core.data/src/org/polymap/core/data/feature/buffer/LayerFeatureBufferManager.java
@@ -317,8 +317,12 @@ public class LayerFeatureBufferManager
             }
             monitor.worked( 1 );
         }
-        fs.addFeatures( added );
-        fs.removeFeatures( ff.id( removed ) );
+        if (!added.isEmpty()) {
+            fs.addFeatures( added );
+        }
+        if (!removed.isEmpty()) {
+            fs.removeFeatures( ff.id( removed ) );
+        }
         
         // none of the features had concurrent modifications, so just upgrade
         // timestamp for the layer (no checking is needed and done)
